Reworking cost structure
